Move jni and asset manager code into platform
